requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
anthropic>=0.40.0
openai>=1.50.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
pyyaml>=6.0.0
